# Improved unique set & vocab methods (reduced runtime 5-10%)