--- a/dataprofiler/profilers/text_column_profile.py
+++ b/dataprofiler/profilers/text_column_profile.py
@@ -2,7 +2,7 @@ from .numerical_column_stats import NumericStatsMixin
 from .base_column_profilers import BaseColumnProfiler, \
     BaseColumnPrimitiveTypeProfiler
 from .profiler_options import TextOptions
-
+import itertools
 
 class TextColumn(NumericStatsMixin, BaseColumnPrimitiveTypeProfiler):
     """
@@ -107,9 +107,8 @@ class TextColumn(NumericStatsMixin, BaseColumnPrimitiveTypeProfiler):
         :type subset_properties: dict
         :return: None
         """
-        vocab = []
-        for row in data:
-            self.vocab = TextColumn._combine_unique_sets(self.vocab, list(row))
+        data_flat = list(itertools.chain(*data))
+        self.vocab = TextColumn._combine_unique_sets(self.vocab, data_flat)
 
     def _update_helper(self, df_series_clean, profile):
         """