--- a/dataprofiler/tests/profilers/test_base_column_profilers.py
+++ b/dataprofiler/tests/profilers/test_base_column_profilers.py
@@ -8,6 +8,7 @@ from unittest.mock import patch, MagicMock
 import pandas as pd
 import numpy as np
 
+from dataprofiler.profilers import utils
 from dataprofiler.tests.profilers import utils as test_utils
 from dataprofiler.profilers.base_column_profilers import BaseColumnProfiler, \
     BaseColumnPrimitiveTypeProfiler
@@ -158,7 +159,7 @@ class TestBaseColumnPrimitiveTypeProfileClass(unittest.TestCase):
     def test_combine_unqiue_sets(self):
         a = [1, 2, 3]
         b = [3, 1, 4, -1]
-        c = BaseColumnPrimitiveTypeProfiler._combine_unique_sets(a, b)
+        c = utils._combine_unique_sets(a, b)
         six.assertCountEqual(self, [1, 2, 3, 4, -1], c)
 
     def test__init__(self):