Generated 1.35m integers and compared the old to new method.

The new method is 3x faster. If order is desired on output, we can sort on report.

**This PR** (new method)
```
5 function calls in 0.053 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.050    0.050    0.050    0.050 test_dp.py:17(_combine_unique_sets_v2)
        1    0.003    0.003    0.053    0.053 test_dp.py:43(combine_unique_sets_v2)
        1    0.000    0.000    0.000    0.000 {built-in method builtins.print}
        1    0.000    0.000    0.000    0.000 {built-in method builtins.len}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
```

**Current Master** (old method)
```
6 function calls in 0.179 seconds

   Ordered by: internal time

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
        1    0.124    0.124    0.124    0.124 {built-in method fromkeys}
        1    0.052    0.052    0.176    0.176 test_dp.py:7(_combine_unique_sets)
        1    0.003    0.003    0.179    0.179 test_dp.py:39(combine_unique_sets_v1)
        1    0.000    0.000    0.000    0.000 {built-in method builtins.print}
        1    0.000    0.000    0.000    0.000 {built-in method builtins.len}
        1    0.000    0.000    0.000    0.000 {method 'disable' of '_lsprof.Profiler' objects}
```


-------------------------

Vocab method was also updated (no reason to merge sets twice and flatten data). An experiment was ran where everything was updated (besides vocab) and you can see the results below using the old vs new method.

**New vocab method update, new set update method**
```
225184188 function calls (208307436 primitive calls) in 112.562 seconds

   Ordered by: internal time
   List reduced from 1393 to 50 due to restriction <50>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
3370/2666   12.561    0.004   12.565    0.005 {built-in method numpy.array}
 16716662    8.738    0.000    8.739    0.000 {method 'sub' of 're.Pattern' objects}
 16716402    6.861    0.000   34.854    0.000 replace.py:122(re_replacer)
50717956/34001547    6.364    0.000   17.153    0.000 {built-in method builtins.isinstance}
       12    5.494    0.458   11.459    0.955 {pandas._libs.lib.map_infer_mask}
 16716409    5.339    0.000    7.836    0.000 typing.py:713(__subclasscheck__)
 16716402    4.753    0.000    4.753    0.000 numerical_column_stats.py:549(is_int)
 16716409    4.383    0.000   12.219    0.000 typing.py:710(__instancecheck__)
```

**Old vocab method update, new set update method**
```
292049730 function calls (275172990 primitive calls) in 149.756 seconds

   Ordered by: internal time
   List reduced from 1393 to 50 due to restriction <50>

   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
----->  6   13.519    2.253   38.406    6.401 text_column_profile.py:95(_update_vocab)
3370/2666   12.706    0.004   12.710    0.005 {built-in method numpy.array}
33523750/33523732   12.428    0.000   12.428    0.000 numerical_column_stats.py:89(__getattribute__)
 16716662    8.809    0.000    8.809    0.000 {method 'sub' of 're.Pattern' objects}
 16716408    8.123    0.000   12.630    0.000 base_column_profilers.py:49(_combine_unique_sets)
 16716402    6.832    0.000   34.713    0.000 replace.py:122(re_replacer)
50717956/34001547    6.308    0.000   16.968    0.000 {built-in method builtins.isinstance}
       12    5.574    0.465   11.594    0.966 {pandas._libs.lib.map_infer_mask}
```